pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-radar"
version = "0.1.0"
description = "本地运行的 AI Radar / AI Research Intelligence System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "arxiv",
  "requests",
  "feedparser",
  "python-dotenv",
  "PyYAML",
  "rich",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
